Refactor and add search filter for country, city and hostname
