+++ src/components/SearchLocation.vue
@@ -2,7 +2,7 @@
 import { onMounted, ref } from 'vue';
 import { NInput } from 'naive-ui';
 
-import useFilteredSocksProxies from '@/composables/useFilteredSocksProxies';
+import useFilteredSocksProxies from '@/composables/useFilteredSocksProxies/useFilteredSocksProxies';
 
 const { query } = useFilteredSocksProxies();
