+++ src/helpers/getRandomSocksProxy.ts
@@ -1,16 +1,16 @@
-import type { Country } from '@/composables/useSocksProxies';
+import { Country } from '@/composables/useFilteredSocksProxies/SocksProxies.types';
 
 type Props = {
-  socksProxies: Country[] | undefined
+  socksProxies: Country[] | undefined;
   country: string;
   city?: string;
-}
+};
 
 const getRandomSocksProxy = ({ socksProxies, country, city }: Props) => {
   if (!socksProxies || !socksProxies.length) {
     throw new Error('No proxies to choose from');
   }
-  
+
   const theCountry = socksProxies.filter((c) => c.country === country)[0];
   let proxies;
   if (!city) {
@@ -23,7 +23,7 @@ const getRandomSocksProxy = ({ socksProxies, country, city }: Props) => {
   }
   const index = Math.floor(Math.random() * proxies.length);
   const { hostname, port } = proxies[index];
-  
+
   return { hostname, port };
 };
